# docs/backend/api.yaml
openapi: 3.0.3
info:
  title: Paperflow Java Admin API
  version: 0.1.0
  description: REST API over the Paperflow PostgreSQL database. Paperflow business data is read-only; Admin User management writes only admin_user.
servers:
  - url: http://localhost:8080
security:
  - sessionCookie: []
...
paths:
...
  /api/papers:
    post:
      tags: [Papers]
      operationId: createPaper
      summary: Create one Paper and upload one Original File.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [metadata, file]
              properties:
                metadata:
                  $ref: "#/components/schemas/PaperMetadataInput"
                file:
                  type: string
                  format: binary
      responses:
        "201":
          description: Paper created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaperMutation"
        "400":
          $ref: "#/components/responses/ValidationError"
        "409":
          description: Paper already exists or is in trash.
    get:
      tags: [Papers]
      operationId: listPapers
      summary: List Paper management records by Original File.
      description: Default sort is year DESC NULLS LAST, then fileId ASC. The optional q filter searches Original File title, authors, DOI, and file ID.
      parameters:
        - name: q
          in: query
          required: false
          schema:
            type: string
            maxLength: 500
        - name: sourceId
          in: query
          required: false
          schema:
            type: string
            pattern: "^S.+"
        - name: sourceName
          in: query
          required: false
          schema:
            type: string
            maxLength: 200
        - name: provider
          in: query
          required: false
          schema:
            type: string
            maxLength: 200
        - name: yearFrom
          in: query
          required: false
          schema:
            type: integer
        - name: yearTo
          in: query
          required: false
          schema:
            type: integer
        - name: flagMatch
          in: query
          required: false
          schema:
            type: integer
            enum: [-1, 0, 1]
        - name: flagText
          in: query
          required: false
          schema:
            type: integer
            enum: [-2, -1, 0, 1, 2]
        - name: flagBlock
          in: query
          required: false
          schema:
            type: integer
            enum: [-1, 0, 1]
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Size"
      responses:
        "200":
          description: Paper page.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaperPage"
        "400":
          $ref: "#/components/responses/ValidationError"
  /api/papers/export:
    get:
      tags: [Papers]
      operationId: exportPapers
      summary: Export Paper management records as UTF-8 CSV.
      responses:
        "200":
          description: CSV export.
          content:
            text/csv:
              schema:
                type: string
                format: binary
        "400":
          $ref: "#/components/responses/ValidationError"
  /api/papers/{fileId}:
    put:
      tags: [Papers]
      operationId: updatePaper
      summary: Update editable Paper metadata using optimistic locking.
      parameters:
        - $ref: "#/components/parameters/FileId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaperUpdateRequest"
      responses:
        "200":
          description: Paper updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaperMutation"
        "409":
          description: Version conflict or matched metadata is locked.
    delete:
      tags: [Papers]
      operationId: softDeletePaper
      summary: Move a Paper to trash. ADMIN or SUPER_ADMIN only.
      parameters:
        - $ref: "#/components/parameters/FileId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaperDeleteRequest"
      responses:
        "200":
          description: Paper moved to trash.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaperMutation"
        "403":
          $ref: "#/components/responses/Forbidden"
    get:
      tags: [Papers]
      operationId: getPaper
      summary: Get Original File metadata, task status, optional OpenAlex metadata, text files, and causal summary.
      parameters:
        - $ref: "#/components/parameters/FileId"
      responses:
        "200":
          description: Paper detail.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaperDetail"
        "404":
          $ref: "#/components/responses/PaperNotFound"
  /api/papers/trash:
    get:
      tags: [Papers]
      operationId: listTrashedPapers
      summary: List soft-deleted Papers. ADMIN or SUPER_ADMIN only.
      responses:
        "200":
          description: Trashed Paper identities and record versions.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TrashedPaper"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/papers/{fileId}/restore:
    post:
      tags: [Papers]
      operationId: restorePaper
      parameters:
        - $ref: "#/components/parameters/FileId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaperVersionActionRequest"
      responses:
        "200":
          description: Paper restored.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaperMutation"
  /api/papers/{fileId}/purge:
    post:
      tags: [Papers]
      operationId: purgePaper
      summary: Permanently delete a trashed Paper. SUPER_ADMIN only.
      parameters:
        - $ref: "#/components/parameters/FileId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaperPurgeRequest"
      responses:
        "204":
          description: Paper permanently deleted.
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/papers/{fileId}/versions:
    get:
      tags: [Papers]
      operationId: listPaperVersions
      parameters:
        - $ref: "#/components/parameters/FileId"
      responses:
        "200":
          description: Original File versions.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PaperFileVersion"
    post:
      tags: [Papers]
      operationId: replacePaperFile
      parameters:
        - $ref: "#/components/parameters/FileId"
        - name: recordVersion
          in: query
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "200":
          description: New file version created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaperMutation"
  /api/papers/{fileId}/versions/{versionNo}/restore:
    post:
      tags: [Papers]
      operationId: restorePaperVersion
      parameters:
        - $ref: "#/components/parameters/FileId"
        - name: versionNo
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaperVersionActionRequest"
      responses:
        "200":
          description: Historical file restored as a new version.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaperMutation"
...
  /api/papers/{fileId}/blocks:
    get:
      tags: [Papers]
      operationId: listPaperBlocks
      summary: List blocks imported for a Paper's Original File.
      parameters:
        - $ref: "#/components/parameters/FileId"
        - name: includeDiscarded
          in: query
          required: false
          schema:
            type: boolean
            default: false
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Size"
      responses:
        "200":
          description: Block page ordered by blockSeq ASC.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlockPage"
        "404":
          $ref: "#/components/responses/PaperNotFound"
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: JSESSIONID
  parameters:
...
    Page:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
    Size:
      name: size
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
    FileId:
      name: fileId
      in: path
      required: true
      schema:
        type: string
        minLength: 1
...
  responses:
    ValidationError:
      description: Invalid request parameters.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            validation:
              value:
                code: VALIDATION_ERROR
                message: Invalid request
                requestId: 9c0d2d7e3f1a4b61
...
    Forbidden:
      description: The current Admin User does not have permission.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            forbidden:
              value:
                code: FORBIDDEN
                message: Forbidden
                requestId: 9c0d2d7e3f1a4b61
...
    PaperNotFound:
      description: Paper management record not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            paperNotFound:
              value:
                code: PAPER_NOT_FOUND
                message: Paper not found
                requestId: 9c0d2d7e3f1a4b61
...
  schemas:
    ErrorResponse:
      type: object
      required: [code, message, requestId]
      properties:
        code:
          type: string
          enum:
            - VALIDATION_ERROR
            - UNAUTHORIZED
            - FORBIDDEN
            - ADMIN_USER_NOT_FOUND
            - ADMIN_USER_CONFLICT
            - SOURCE_NOT_FOUND
            - WORK_NOT_FOUND
            - ORIGINAL_FILE_NOT_FOUND
            - PAPER_NOT_FOUND
            - PAPER_ALREADY_EXISTS
            - PAPER_IN_TRASH
            - PAPER_VERSION_CONFLICT
            - PAPER_MATCHED_FIELDS_LOCKED
            - PAPER_STORAGE_CONFLICT
            - INVALID_PAPER_FILE
            - PAPER_FILE_TOO_LARGE
            - PURGE_CONFIRMATION_REQUIRED
            - INTERNAL_ERROR
        message:
          type: string
        requestId:
          type: string
...
    WorkMetadata:
      type: object
      required: [workId]
      properties:
        workId:
          type: string
        title:
          type: string
          nullable: true
        doi:
          type: string
          nullable: true
        publicationYear:
          type: integer
          nullable: true
        publicationDate:
          type: string
          nullable: true
        type:
          type: string
          nullable: true
        language:
          type: string
          nullable: true
...
    MatchedFile:
      type: object
      required:
        - fileId
        - sourceId
        - originalFileName
        - originalFilePath
        - originalFileUrl
        - originalFileType
        - fileSize
        - flagMatch
        - flagText
        - flagBlock
        - textFiles
      properties:
        fileId:
          type: string
        sourceId:
          type: string
        sourceName:
          type: string
          nullable: true
        year:
          type: integer
          nullable: true
        paperTitle:
          type: string
          nullable: true
        authors:
          type: string
          nullable: true
        doi:
          type: string
          nullable: true
        url:
          type: string
          nullable: true
        provider:
          type: string
          nullable: true
        originalFileName:
          type: string
        originalFilePath:
          type: string
        originalFileUrl:
          type: string
        originalFileType:
          type: string
          enum: [PDF, XML, HTML]
        fileSize:
          type: integer
          format: int64
        recordVersion:
          type: integer
          format: int64
        currentVersion:
          type: integer
        createdAt:
          type: string
          format: date-time
          nullable: true
        createdBy:
          type: integer
          format: int64
          nullable: true
        updatedAt:
          type: string
          format: date-time
          nullable: true
        updatedBy:
          type: integer
          format: int64
          nullable: true
        flagMatch:
          type: integer
        matchedWorkId:
          type: string
          nullable: true
        flagText:
          type: integer
        flagBlock:
          type: integer
        textFiles:
          type: array
          items:
            $ref: "#/components/schemas/TextFile"
    OriginalFilePage:
      type: object
      required: [items, page, size, total]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/OriginalFile"
        page:
          type: integer
        size:
          type: integer
        total:
          type: integer
          format: int64
    PaperPage:
      allOf:
        - $ref: "#/components/schemas/OriginalFilePage"
    OriginalFile:
      allOf:
        - $ref: "#/components/schemas/MatchedFile"
    PaperDetail:
      type: object
      required: [originalFile, taskStatus, textFiles]
      properties:
        originalFile:
          $ref: "#/components/schemas/OriginalFile"
        taskStatus:
          type: object
          required: [flagMatch, flagText, flagBlock]
          properties:
            flagMatch:
              type: integer
            flagText:
              type: integer
            flagBlock:
              type: integer
        openAlex:
          nullable: true
          allOf:
            - $ref: "#/components/schemas/WorkMetadata"
        textFiles:
          type: array
          items:
            $ref: "#/components/schemas/TextFile"
        causalSummary:
          nullable: true
          allOf:
            - $ref: "#/components/schemas/CausalPaperSummary"
    PaperMetadataInput:
      type: object
      required: [sourceId, year, paperTitle, authors]
      properties:
        sourceId:
          type: string
        year:
          type: integer
          minimum: 1000
        paperTitle:
          type: string
          maxLength: 2000
        authors:
          type: array
          minItems: 1
          items:
            type: string
            maxLength: 500
        doi:
          type: string
          nullable: true
        url:
          type: string
          nullable: true
    PaperUpdateRequest:
      allOf:
        - $ref: "#/components/schemas/PaperMetadataInput"
        - type: object
          required: [recordVersion]
          properties:
            recordVersion:
              type: integer
              format: int64
              minimum: 0
    PaperMutation:
      type: object
      required: [fileId, recordVersion]
      properties:
        fileId:
          type: string
        recordVersion:
          type: integer
          format: int64
    TrashedPaper:
      allOf:
        - $ref: "#/components/schemas/PaperMutation"
        - type: object
          required: [sourceId, recordVersion, deletedAt]
          properties:
            sourceId:
              type: string
            sourceName:
              type: string
              nullable: true
            year:
              type: integer
              nullable: true
            paperTitle:
              type: string
              nullable: true
            authors:
              type: string
              nullable: true
            deletedAt:
              type: string
              format: date-time
            deletedBy:
              type: integer
              format: int64
              nullable: true
            deleteReason:
              type: string
              nullable: true
    PaperDeleteRequest:
      type: object
      required: [recordVersion]
      properties:
        recordVersion:
          type: integer
          format: int64
        reason:
          type: string
          maxLength: 500
          nullable: true
    PaperVersionActionRequest:
      type: object
      required: [recordVersion]
      properties:
        recordVersion:
          type: integer
          format: int64
    PaperPurgeRequest:
      type: object
      required: [confirmation, recordVersion]
      properties:
        confirmation:
          type: string
          enum: [删除]
        recordVersion:
          type: integer
          format: int64
    PaperFileVersion:
      type: object
      required: [fileId, versionNo, fileName, fileUrl, fileType, fileSize, uploadedAt, current]
      properties:
        fileId:
          type: string
        versionNo:
          type: integer
        fileName:
          type: string
        fileUrl:
          type: string
        fileType:
          type: string
          enum: [PDF, XML, HTML]
        fileSize:
          type: integer
          format: int64
        uploadedBy:
          type: integer
          format: int64
          nullable: true
        uploadedAt:
          type: string
          format: date-time
        current:
          type: boolean
...
    TextFile:
      type: object
      required: [fileId, fileType, fileName, filePath, fileUrl, fileSize]
      properties:
        fileId:
          type: string
        fileType:
          type: string
          enum: [JSON, MD]
        fileName:
          type: string
        filePath:
          type: string
        fileUrl:
          type: string
        fileSize:
          type: integer
          format: int64
    BlockPage:
      type: object
      required: [items, page, size, total]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Block"
        page:
          type: integer
        size:
          type: integer
        total:
          type: integer
          format: int64
    Block:
      type: object
      required: [blockId, fileId, blockType, blockSeq]
      properties:
        blockId:
          type: string
        fileId:
          type: string
        blockType:
          type: string
          enum: [title, text, equation, table, image, reference, page_footnote, discarded]
        blockText:
          type: string
          nullable: true
        pdfPage:
          type: integer
          nullable: true
        pdfBbox:
          nullable: true
        blockSeq:
          type: integer
        parentTitleBlockId:
          type: string
          nullable: true
        titleLevel:
          type: integer
          nullable: true
        imagePath:
          type: string
          nullable: true
        imageUrl:
          type: string
          nullable: true
        imageCaption:
          type: string
          nullable: true
        imageFootnote:
          type: string
          nullable: true
        tableImagePath:
          type: string
          nullable: true
        tableImageUrl:
          type: string
          nullable: true
        tableCaption:
          type: string
          nullable: true
        tableFootnote:
          type: string
          nullable: true
        equationImagePath:
          type: string
          nullable: true
        equationImageUrl:
          type: string
          nullable: true
        equationFormat:
          type: string
          nullable: true
        footnoteLabel:
          type: string
          nullable: true
        footnoteText:
          type: string
          nullable: true
        references:
          type: array
          items:
            type: string
          nullable: true
...
    CausalPaperSummary:
      type: object
      required: [workId, claimRecordCount, standardClaimCount, variableCount, hasCausalClaims]
      properties:
        workId:
          type: string
        claimRecordCount:
          type: integer
          format: int64
        standardClaimCount:
          type: integer
          format: int64
        variableCount:
          type: integer
          format: int64
        hasCausalClaims:
          type: boolean
...